+++ packages/expo-facebook/ios/EXFacebook/EXFacebook.m
@@ -32,21 +32,6 @@ UM_EXPORT_METHOD_AS(setAutoLogAppEventsEnabledAsync,
   resolve(nil);
 }
 
-UM_EXPORT_METHOD_AS(setAutoInitEnabledAsync,
-                    setAutoInitEnabled:(BOOL)enabled
-                    resolver:(UMPromiseResolveBlock)resolve
-                    rejecter:(UMPromiseRejectBlock)reject)
-{
-  // If enabled is true, the line below will initialize the SDK.
-  // This behavior is different than on Android where one needs
-  // to initialize the SDK explicitly. We have no power over this,
-  // and to mitigate this difference we will NOT add initializing
-  // to the respective method on Android, but we will instruct users
-  // to initialize the SDK manually on both platforms instead.
-  // [FBSDKSettings setAutoInitEnabled:enabled];
-  resolve(nil);
-}
-
 UM_EXPORT_METHOD_AS(initializeAsync,
                     initializeAsync:(NSDictionary *)options
                     resolver:(UMPromiseResolveBlock)resolve